U/d pickOrange call signature to take a tree arg
These tests invoke a method called `dropOrange()`
which is a global function.

In all the other tests we are required to
initialize a tree "instance" before running the
test.

The `dropOrange` tests do not do this.

Given that the tree variables are locally defined,
it is _not_ possible for `dropOrange` to access
the tree and pull an `Orange` out of it.

A sensible call pattern would be `dropOrange(tree)`
(provided we defined `tree`). This would illustrate
that it is possible to define global functions
that operate on (pseudo-)instances.