DRY out, Improve readibility and prompting in spec
There were a number of calls to the same logic
that was repeating the initialization of the stub
objects.

ALSO there was repetition in several places of the
aging mechanism.

ALSO there were tests that assumed that the tree
in question was an aged tree.

The test structure should now reveal these
implicit dependencies.